NEEEEWWWW

+++ week 5/day 1/C_paint_the_array.cpp
@@ -19,7 +19,7 @@ int main()
     {
         /* code */
         long long n;     cin>>n;
-        vector<int>a(n);
+        vector<long long>a(n);
         for(int i=0;i<n;i++){
             cin>>a[i];
         }
